✨ (accounts): add modularized account management system
Refactor the account management system by organizing it into modules
for categories, certifications, onboardings, service images, services,
sessions, subcategories, user profiles, and users. This modularization
improves code maintainability and scalability by separating concerns
and allowing for easier future enhancements.
