+++ server/cmd/main.go
@@ -3,15 +3,15 @@ package main
 import (
 	"conecta-mare-server/internal/config"
 	"conecta-mare-server/internal/database"
-	"conecta-mare-server/internal/modules/categories"
-	"conecta-mare-server/internal/modules/certifications"
-	"conecta-mare-server/internal/modules/onboardings"
-	"conecta-mare-server/internal/modules/serviceimages"
-	"conecta-mare-server/internal/modules/services"
-	"conecta-mare-server/internal/modules/session"
-	"conecta-mare-server/internal/modules/subcategories"
-	"conecta-mare-server/internal/modules/userprofiles"
-	"conecta-mare-server/internal/modules/users"
+	"conecta-mare-server/internal/modules/accounts/categories"
+	"conecta-mare-server/internal/modules/accounts/certifications"
+	"conecta-mare-server/internal/modules/accounts/onboardings"
+	"conecta-mare-server/internal/modules/accounts/serviceimages"
+	"conecta-mare-server/internal/modules/accounts/services"
+	"conecta-mare-server/internal/modules/accounts/session"
+	"conecta-mare-server/internal/modules/accounts/subcategories"
+	"conecta-mare-server/internal/modules/accounts/userprofiles"
+	"conecta-mare-server/internal/modules/accounts/users"
 	"conecta-mare-server/internal/server"
 	"conecta-mare-server/pkg/jwt"
 	"conecta-mare-server/pkg/storage"
