+++ server/internal/modules/accounts/categories/entity.go
