+++ server/internal/modules/accounts/serviceimages/interface.go
